input has sounds. favicon added
