+++ app/static/js/documentElements-game.js
@@ -22,9 +22,3 @@ export const gameEndPanel = document.getElementById('game-end-panel');
 export const playAgainButton = gameEndPanel.querySelector('#play-again-button');
 export const reviewMapButton = gameEndPanel.querySelector('#review-button');
 export const homeButton = gameEndPanel.querySelector('#home-button');
-
-export const SVG_WIDTH = 1600;
-export const SVG_HEIGHT = 900;
-export const ATTEMPT_COLORS = [ 'rgb(75, 255, 75)', 'rgb(240, 219, 35)', 'rgb(243, 148, 24)', 'rgb(235, 89, 89)' ];
-export const REPEAT_COLOR = "rgba(172, 233, 164, 1)";
-export const MAX_GUESSES = 3;
