+++ app/static/js/variables.js
@@ -1,0 +1,9 @@
+export const SVG_WIDTH = 1600;
+export const SVG_HEIGHT = 900;
+export const SVG_PADDING = 20; // pixels
+
+export const ATTEMPT_COLORS = [ 'rgb(75, 255, 75)', 'rgb(240, 219, 35)', 'rgb(243, 148, 24)', 'rgb(235, 89, 89)' ];
+export const REPEAT_COLOR = "rgba(172, 233, 164, 1)";
+export const MAX_GUESSES = 3;
+
+export const ATTEMPT_SOUNDS = [ "correct.wav", "neutral.wav", "neutral.wav", "incorrect.wav", "incorrectPlus.wav" ];
