+++ app/templates/game.html
@@ -5,6 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
+    <link rel="shortcut icon" href="images/favicon.ico" type="image/x-icon">
+
     <link rel="stylesheet" href="css/all.css" />
     <link rel="stylesheet" href="css/header.css" />
     <link rel="stylesheet" href="css/game.css"/>
